// itmcalibraciones-backend/src/errors-handler/enums/status.enum.ts
export enum StatusEnum {
  OK = 200,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  INTERNAL_SERVER_ERROR = 500,
  USER_ALREADY_EXISTS = 1000,
  USER_NOT_FOUND = 1001,
  INSTRUMENT_NOT_BELONG_TO_USER = 1002,
  EMPLOYEE_NOT_FOUND = 1003,
  COMPANY_NOT_FOUND = 1004,
  LOGIN_USER_NOT_FOUND = 1005,
  LOGIN_USER_FAIL = 1006,
  PASSWORD_INCORRECT = 1007,
  FIELD_NOT_EXIST = 1008,
  CLIENT_NOT_EXIST = 1009,
  CITY_NOT_FOUND = 1010,
  OFFICE_NOT_FOUND = 1011,
  EVENT_NOT_RELATED_TO_USER = 1012,
  EVENT_SCORE_NOT_PROCESS_IS_OPEN = 1013,
  EVENT_IS_CLOSE = 1014,
  CROSS_NOT_FOUND = 1015,
  OPTION_NOT_RELATED_TO_FIELD = 1016,
  CROSS_IS_CLOSE = 1017,
  EVENT_ALREADY_ADDED = 1018,
  TOKEN_USED = 1019,
  USER_NOT_VERIFY = 1020,
  AVOID_SEND_EMAIL_TO_REAL_USER_ON_DEV = 1021,
  ERROR_RECAPTCHA_NOT_VALID = 1022,
  ERROR_VERYFING_RECAPTCHA = 1023,
  ERROR_DELETE_CERTIFICATE_UNDER_USING = 1024
}